+++ lib/launcher_helper.dart
@@ -12,3 +12,4 @@ export 'src/_launcher_helper.dart';
 export 'src/palette_generator.dart';
 export 'src/application_collection.dart';
 export 'src/_icon.dart';
+export 'src/_icon_shape.dart';